+++ src/main/java/com/fifo/ticketing/domain/book/repository/BookScheduleRepository.java
@@ -1,0 +1,10 @@
+package com.fifo.ticketing.domain.book.repository;
+
+import com.fifo.ticketing.domain.book.entity.BookScheduledTask;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookScheduleRepository extends JpaRepository<BookScheduledTask, Long> {
+
+}
