+++ src/main/java/com/fifo/ticketing/domain/book/service/BookScheduleService.java
@@ -1,0 +1,32 @@
+package com.fifo.ticketing.domain.book.service;
+
+import com.fifo.ticketing.domain.book.mapper.BookMapper;
+import com.fifo.ticketing.domain.book.repository.BookScheduleRepository;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Date;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.scheduling.TaskScheduler;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class BookScheduleService {
+
+    private final BookService bookService;
+    private final BookScheduleRepository bookScheduleRepository;
+    @Qualifier("taskScheduler")
+    private final TaskScheduler taskScheduler;
+
+    public void scheduleCancelTask(Long bookId, LocalDateTime runTime) {
+
+        bookScheduleRepository.save(BookMapper.toBookScheduledTaskEntity(bookId, runTime));
+
+        Date triggerTime = Date.from(runTime.atZone(ZoneId.systemDefault()).toInstant());
+
+        taskScheduler.schedule(() -> bookService.cancelIfUnpaid(bookId), triggerTime);
+
+    }
+
+}
